Add feature to support email-less omniauth backends
If backend do not provide auth.info.email, We generate it from auth.info.nickname and email_domain option passed to omniauth provider (if it's passed).

You should be able to pass this option to all omniauth providers (trusted and normal).

This feature is usefull in environments where login is done for example by LDAP or pam and that backend is not providing email. (ldap is missing email field. pam can't normally know it. etc..)